Refactor for reduced duplication
Add bullet points for solutions list

Add summary for report

Add visual separation for Uncategorized issues
